mostly added cout comments for explanation of steps

+++ OpenGL/src/test/Test.h
@@ -12,7 +12,7 @@ namespace test {
 		Test() {}
 		virtual ~Test() {}
 
-		virtual void onUpdate(float deltaTime) {}
+		virtual void onUpdate(double deltaTime) {}
 		virtual void onRender() {}
 		virtual void onImGuiRender() {}
